external/DAPA_release/requirements.txt
numpy
opencv-python
pyopengl
pyrender
scikit-image
scipy==1.0.0
tensorboard
chumpy
smplx==0.1.13
#torch==1.1.0
torchgeometry
#torchvision
tqdm
trimesh
neural-renderer-pytorch==1.1.3
progress
dotmap
wandb
ipdb
# human-body-prior==2.1.2.0  need to run python setup.py install from their repo.
